#0006630 ESDEV-4460 Reset session basket on module deactivation
It is an eShop default behaviour to store the current basket object in the session.
If the PayPal module is active, the current basket object is an instance of a \OxidEsales\PayPalModule\Model\Basket.
After module deactivation the parent class of this class is no longer available and the object deserialization fails with a fatal error.
As fix for bug #0006630 the current basket is deleted from the session on module deactivation.