Fix compatibility with other modules
Due to the fact that there is no query builder API present, the sql string for admin select lists is manipulated directly.
A lot of modules use `str_replace` to find the part `from oxorder` in order to add further columns or joins to the query.

However, this paypal module breaks compatiblity between modules by replacing `from oxorder` with the backticks-quoted identifier variant.
This can lead to exceptions being thrown when users click on a sorting link in the list's header (because for example some some joins could not be applied with `str_replace`).

This PR replaces the quoted oxorder identifier with its unquoted variant. This variant is also used by in the sql string returned from the parent class (`OrderList::_buildSelectString()`)).